% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim}
\alias{sim}
\title{Runs a portfolio simulation in C++}
\usage{
sim(portfolio, n_factor, n_sim)
}
\arguments{
\item{portfolio}{\code{matrix} porfolio}

\item{n_factor}{\code{integer} number of factors}

\item{n_sim}{\code{integer} number of simulations}
}
\description{
Runs a portfolio simulation in C++
}
