% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_xorshift.R
\name{simulate_xorshift}
\alias{simulate_xorshift}
\title{Runs a portfolio simulation from a GPU}
\usage{
simulate_xorshift(N = 1e+06)
}
\arguments{
\item{N}{\code{numeric} sample size}
}
\description{
Runs a portfolio simulation from a GPU
}
