% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulator}
\alias{simulator}
\title{Runs GPU-based simulator}
\usage{
simulator(portfolio, n_factor, n_sim)
}
\arguments{
\item{portfolio}{\code{matrix} porfolio}

\item{n_factor}{\code{integer} number of factors}

\item{n_sim}{\code{integer} number of simulations}
}
\description{
Runs GPU-based simulator
}
