% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xorshift_generator}
\alias{xorshift_generator}
\title{Runs GPU-based simulator}
\usage{
xorshift_generator(numbers, N)
}
\arguments{
\item{numbers}{\code{vector} input vector}

\item{N}{\code{integer} number of samples}
}
\description{
Runs GPU-based simulator
}
